requirements.txt
fastapi==0.98.0
pydantic==1.10.10
httpx
uvicorn
aiofiles
redis-om
redis
python-jose
numpy
pandas
requests
psycopg2-binary
